requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pyodbc==5.2.0
gunicorn==20.0.4
python-dotenv==1.0.0
